Speed up execution by not using find() on a two-Dimensional Array
Best practice is to separate output variables with comma.

To speed up and simplify the program, replace the call to find with logical array indexing. If the logical array is not a vector, the shape of the output could change when you remove (or use Code Analyzer automatic fix feature to remove) the find call . To retain the same output shape, you might need to replace find(L) by L(:) or L(:)'. Even with these changes, it is faster than calling find.